backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
pydantic-settings==2.2.1
python-jose[cryptography]==3.3.0
PyJWT==2.8.0
httpx==0.27.0
gitpython==3.1.43
pyyaml==6.0.1
python-multipart==0.0.9
